Update ProgressBar integers and includes

+++ src/inc/ProgressBar.h
@@ -1,15 +1,15 @@
 #include <iostream>
-#include <cstdint>
+#include <stdint.h>
 
 void ShowProgressBar(uint64_t& total, uint64_t& cmpl){
-	int fill = 32 * cmpl / total;
+	uint8_t fill = 32 * cmpl / total;
 	std::cout << "\r[";
-	for (int i = 0; i < fill; i++) {
+	for (uint8_t i = 0; i < fill; i++) {
 		std::cout << "=";
 	}
-	for (int i = 0; i < 32 - fill; i++) {
+	for (uint8_t i = 0; i < 32 - fill; i++) {
 		std::cout << " ";
 	}
-	int percent = 100 * cmpl / total;
-	std::cout << "] " << percent << "%" << std::flush;
+	uint8_t percent = 100 * cmpl / total;
+	std::cout << "] " << static_cast<int>(percent) << "%" << std::flush;
 }
